split out some helper types
