+++ src/o1js-missing.ts
@@ -1,0 +1,28 @@
+/**
+ * This file exports types and functions that actually should be exported from o1js
+ */
+import { InferProvable, Provable, ProvablePure } from 'o1js';
+
+export { ProvableType, ProvablePureType, InferProvableType };
+
+const ProvableType = {
+  get<A extends WithProvable<any>>(type: A): ToProvable<A> {
+    return (
+      (typeof type === 'object' || typeof type === 'function') &&
+      type !== null &&
+      'provable' in type
+        ? type.provable
+        : type
+    ) as ToProvable<A>;
+  },
+};
+
+type WithProvable<A> = { provable: A } | A;
+type ProvableType<T = any, V = any> = WithProvable<Provable<T, V>>;
+type ProvablePureType<T = any, V = any> = WithProvable<ProvablePure<T, V>>;
+type ToProvable<A extends WithProvable<any>> = A extends {
+  provable: infer P;
+}
+  ? P
+  : A;
+type InferProvableType<T extends ProvableType> = InferProvable<ToProvable<T>>;
